requirements.txt
pytest==8.0.0
selenium==4.15.0
pytest-html==4.1.1
allure-pytest==2.13.2
python-dotenv==1.0.0
openpyxl==3.1.2
webdriver-manager==4.0.1
pytest-xdist==3.5.0
